+++ dragonchain/webserver/routes/smart_contracts.py
@@ -35,8 +35,7 @@ def apply_routes(app: flask.Flask):
     app.add_url_rule("/v1/contract", "list_contract_v1", list_contract_v1, methods=["GET"])
     app.add_url_rule("/contract/<contract_id>", "get_contract_by_id_v1", get_contract_by_id_v1, methods=["GET"])
     app.add_url_rule("/v1/contract/<contract_id>", "get_contract_by_id_v1", get_contract_by_id_v1, methods=["GET"])
-    app.add_url_rule("/contract/logs/<contract_id>", "get_smart_contract_logs_v1", get_smart_contract_logs_v1, methods=["GET"])
-    app.add_url_rule("/v1/contract/logs/<contract_id>", "get_smart_contract_logs_v1", get_smart_contract_logs_v1, methods=["GET"])
+    app.add_url_rule("/v1/contract/<contract_id>/logs", "get_smart_contract_logs_v1", get_smart_contract_logs_v1, methods=["GET"])
     app.add_url_rule("/contract/txn_type/<txn_type>", "get_contract_by_txn_type_v1", get_contract_by_txn_type_v1, methods=["GET"])
     app.add_url_rule("/v1/contract/txn_type/<txn_type>", "get_contract_by_txn_type_v1", get_contract_by_txn_type_v1, methods=["GET"])
     app.add_url_rule("/contract", "post_contract_v1", post_contract_v1, methods=["POST"])
